+++ lib/gumdrop/sprockets_support.rb
@@ -1,3 +1,0 @@
-require 'sprockets'
-
-# Any sprockets extentions/helpers